+++ CHANGELOG.md
@@ -8,6 +8,11 @@
   - branch name
   - branch id
   - todos
+- Add `import`/`export` functionality to import or export branches as `.json`
+- Add `move` functionality to move todos between branches
+
+### 🛠️ Fixes
+
 - Rename `[todo].done` to `[todo].completed`
 
 ## `v1.0.0`